+++ custom_components/zaptec/button.py
@@ -26,6 +26,15 @@ _LOGGER = logging.getLogger(__name__)
 class ZaptecButton(ZaptecBaseEntity, ButtonEntity):
     zaptec_obj: Charger
 
+    @property
+    def available(self) -> bool:
+        """Return True if entity is available."""
+        if not self._attr_available:
+            return False
+        else:
+            # Disable/gray out button if the command is invalid in the current state
+            return self.zaptec_obj.is_command_valid(self.key)
+
     async def async_press(self) -> None:
         """Press the button."""
         _LOGGER.debug(
@@ -48,8 +57,6 @@ class ZapButtonEntityDescription(ButtonEntityDescription):
 
 
 INSTALLATION_ENTITIES: list[EntityDescription] = []
-
-CIRCUIT_ENTITIES: list[EntityDescription] = []
 
 CHARGER_ENTITIES: list[EntityDescription] = [
     ZapButtonEntityDescription(
@@ -97,7 +104,6 @@ async def async_setup_entry(
     entities = ZaptecButton.create_from_zaptec(
         coordinator,
         INSTALLATION_ENTITIES,
-        CIRCUIT_ENTITIES,
         CHARGER_ENTITIES,
     )
     async_add_entities(entities, True)