+++ custom_components/zaptec/validate.py
@@ -57,12 +57,6 @@ class Circuit(BaseModel):
     model_config = ConfigDict(extra="allow")
     Id: str
     Name: str
-
-
-class CircuitHierarchy(BaseModel):
-    model_config = ConfigDict(extra="allow")
-    Id: str
-    Name: str
     Chargers: list[Charger]
 
 
@@ -70,7 +64,7 @@ class Hierarchy(BaseModel):
     model_config = ConfigDict(extra="allow")
     Id: str
     Name: str
-    Circuits: list[CircuitHierarchy]
+    Circuits: list[Circuit]
 
 
 class ChargerFirmware(BaseModel):
@@ -114,7 +108,6 @@ URLS = {
     r"installation/[0-9a-f\-]+/hierarchy": Hierarchy,
     r"installation/[0-9a-f\-]+/update": None,
     r"installation/[0-9a-f\-]+/messagingConnectionDetails": InstallationConnectionDetails,
-    r"circuits/[0-9a-f\-]+": Circuit,
     r"chargers/[0-9a-f\-]+": Charger,
     r"chargers/[0-9a-f\-]+/state": ChargerStates,
     r"chargers/[0-9a-f\-]+/authorizecharge": None,